+++ src/main/java/com/kakao/cafe/domain/UserInformation.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe.dto;
+package com.kakao.cafe.domain;
 
 public class UserInformation {
