+++ src/main/java/com/kakao/cafe/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.kakao.cafe.repository;
 
-import com.kakao.cafe.dto.UserInformation;
+import com.kakao.cafe.domain.UserInformation;
 
 import java.util.List;
 import java.util.Optional;
