+++ src/test/java/com/kakao/cafe/repository/MemoryArticleRepositoryTest.java
@@ -2,7 +2,7 @@ package com.kakao.cafe.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.kakao.cafe.dto.UserArticle;
+import com.kakao.cafe.domain.UserArticle;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
